Clean up, remove unused methods
