--- test/blocks/listgroup/andListgroup.html
@@ -1,18 +1,0 @@
-<header></header>
-<main>
-  <div class="section listgroup-container" data-section-status="loaded" style>
-    <div class="listgroup-wrapper">
-      <div class="listgroup">
-        <div>
-          <div>options</div>
-          <div>limit=10,filterType=and</div>
-        </div>
-        <div>
-          <div>industry</div>
-          <div>biotechnology, chemistry</div>
-        </div>
-      </div>
-    </div>
-  </div>
-</main>
-<footer></footer>
