--- test/blocks/listgroup/optionsListgroup.html
@@ -1,18 +1,0 @@
-<header></header>
-<main>
-  <div class="section listgroup-container" data-section-status="loaded" style>
-    <div class="listgroup-wrapper">
-      <div class="listgroup">
-        <div>
-          <div>options</div>
-          <div>limit=6,columns=3,images=off</div>
-        </div>
-        <div>
-          <div>resourceType</div>
-          <div>interview</div>
-        </div>
-      </div>
-    </div>
-  </div>
-</main>
-<footer></footer>
